Remove load

+++ app/Http/Controllers/DriverController.php
@@ -822,8 +822,11 @@ class DriverController extends Controller
 
                     $load->driverCallCounter--;
                     $load->save();
-                    if ($load->driverCallCounter <= 0) {
-                        $load->delete();
+                    $fleets = json_decode($load->fleets, true);
+                    foreach ($fleets as $fleet) {
+                        if ($fleet['fleet_id'] == 82 && $load->driverCallCounter <= 0) {
+                            $load->delete();
+                        }
                     }
                 }
